move create tests to their own file
